+++ src/index.ts
@@ -17,6 +17,8 @@ import {
   SessionToken,
   StopAudioPlayback,
   User,
+  UserProfile,
+  UserProfileField,
 } from './common/data_structures';
 import {
   CHAT_HISTORY_TYPE,
@@ -87,4 +89,6 @@ export {
   TriggerEvent,
   TriggerParameter,
   User,
+  UserProfile,
+  UserProfileField,
 };