+++ src/services/world_engine.service.ts
@@ -1,14 +1,11 @@
 import * as fm from '../../proto/fetch.pb';
-import {
-  ClientRequest,
-  UserRequest,
-  WorldEngine,
-} from '../../proto/world-engine.pb';
+import { ClientRequest, WorldEngine } from '../../proto/world-engine.pb';
 import { CLIENT_ID } from '../common/constants';
 import {
   ExtensionLoadSceneProps,
   InternalClientConfiguration,
   SessionToken,
+  User,
 } from '../common/data_structures';
 
 type PbFunc<P, R> = (req: P, initReq?: fm.InitReq) => Promise<R>;
@@ -16,7 +13,7 @@ type PbFunc<P, R> = (req: P, initReq?: fm.InitReq) => Promise<R>;
 export interface LoadSceneProps {
   name: string;
   client?: ClientRequest;
-  user?: UserRequest;
+  user?: User;
   config: InternalClientConfiguration;
   session: SessionToken;
   sceneProps?: ExtensionLoadSceneProps;
@@ -26,6 +23,7 @@ export class WorldEngineService {
   async loadScene(props: LoadSceneProps) {
     const { client, config, name, sceneProps, session, user } = props;
     const { hostname, ssl } = config.connection.gateway;
+    const { id, fullName, profile } = user;
 
     const url = `${ssl ? 'https' : 'http'}://${hostname}`;
 
@@ -35,8 +33,20 @@ export class WorldEngineService {
         ...client,
       },
       name,
-      user,
+      user: {
+        ...(id && { id }),
+        ...(fullName && { name: fullName }),
+      },
       capabilities: config.capabilities,
+      ...(profile?.fields.length && {
+        userSettings: {
+          playerProfile: {
+            fields: profile.fields.map(
+              ({ id: fieldId, value: fieldValue }) => ({ fieldId, fieldValue }),
+            ),
+          },
+        },
+      }),
       ...sceneProps,
     });
   }